MODINVOICE-59 Prorated adjustments part 2 - validation/special cases
- moved logic related to adjustments to AdjustmentService classes
