+++ src/main/java/org/folio/invoices/utils/ErrorCodes.java
@@ -32,7 +32,10 @@ public enum ErrorCodes {
   INVOICE_PAYMENT_FAILURE("invoicePaymentFailure", "Invoice payment failure"),
   CURRENT_FISCAL_YEAR_NOT_FOUND("currentFYearNotFound", "Current fiscal year not found for ledger"),
   TRANSACTION_CREATION_FAILURE("transactionCreationFailure", "One or more transactions record(s) failed to be created"),
-  DOCUMENT_IS_TOO_LARGE("documentIsTooLarge", "Document size is too large");
+  DOCUMENT_IS_TOO_LARGE("documentIsTooLarge", "Document size is too large"),
+  ADJUSTMENT_IDS_NOT_UNIQUE("adjustmentIdsNotUnique", "Adjustment ids must be unique"),
+  CANNOT_DELETE_ADJUSTMENTS("cannotDeleteAdjustment", "Prorated adjustment cannot be deleted because it is present on invoice level"),
+  CANNOT_ADD_ADJUSTMENTS("cannotAddAdjustment", "Prorated adjustment cannot be adde because it is not present on invoice level");
 
   private final String code;
   private final String description;
